# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Web-Karir (Career Hub) Secure REST API",
        "description": "Dokumentasi REST API lengkap dan terproteksi untuk platform Web-Karir (Career Hub). Menggunakan Laravel Sanctum Bearer Token Authentication, Rate Limiting, dan Spatie Role-Based Access Control.",
        "contact": {
            "name": "Web-Karir Support Team",
            "email": "admin@webkarir.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Web-Karir Primary API Server"
        }
    ],
    "paths": {
...
        "/jobs/{id}/apply": {
            "post": {
                "tags": [
                    "Candidate Applications"
                ],
                "summary": "Mengirim Lamaran Kerja",
                "description": "Pelamar mengirimkan lamaran untuk lowongan pekerjaan tertentu.",
                "operationId": "7ca7333ffe2dc7ab3aaac0c0afef1546",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "cover_letter": {
                                        "type": "string",
                                        "example": "Saya berminat bergabung di posisi ini..."
                                    },
                                    "resume": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lamaran berhasil dikirim",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Lamaran pekerjaan berhasil dikirim!",
                                    "data": {
                                        "id": 10,
                                        "job_id": 1,
                                        "user_id": 1,
                                        "cover_letter": "Saya berminat bergabung...",
                                        "status": "pending",
                                        "created_at": "2026-08-21T22:00:00.000000Z"
                                    },
                                    "errors": null
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sudah melamar atau kuota penuh",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Anda sudah melamar pekerjaan ini sebelumnya.",
                                    "data": null,
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/candidate/applications": {
            "get": {
                "tags": [
                    "Candidate Applications"
                ],
                "summary": "Daftar Riwayat Lamaran Saya",
                "description": "Mengambil riwayat lamaran yang pernah dikirimkan.",
                "operationId": "c557b33a6cec9ad9b0bbe6b8c02f5628",
                "responses": {
                    "200": {
                        "description": "Daftar lamaran kandidat",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Daftar riwayat lamaran berhasil diambil.",
                                    "data": [
                                        {
                                            "id": 10,
                                            "job_id": 1,
                                            "status": "interview",
                                            "created_at": "2026-08-21T22:00:00.000000Z",
                                            "job": {
                                                "title": "Senior Backend Developer",
                                                "company_name": "PT Tech Nusantara"
                                            },
                                            "interviews": [
                                                {
                                                    "interview_date": "2026-09-01 10:00:00",
                                                    "location_or_link": "Google Meet: https://meet.google.com/abc-defg-hij"
                                                }
                                            ]
                                        }
                                    ],
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications/{id}/messages": {
            "get": {
                "tags": [
                    "Candidate Applications"
                ],
                "summary": "Pesan Live Chat Lamaran",
                "description": "Mengambil histori percakapan antara Pelamar dan HR.",
                "operationId": "a1520c55e78dabae4cad1423283be014",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histori percakapan chat",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Pesan percakapan berhasil diambil.",
                                    "data": [
                                        {
                                            "id": 1,
                                            "application_id": 10,
                                            "sender_id": 1,
                                            "message": "Halo HR, mohon info jadwal wawancara",
                                            "created_at": "2026-08-21T22:05:00.000000Z",
                                            "sender": {
                                                "id": 1,
                                                "name": "Budi Pratama"
                                            }
                                        }
                                    ],
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Candidate Applications"
                ],
                "summary": "Kirim Pesan Live Chat",
                "description": "Mengirim pesan baru pada saluran obrolan lamaran kerja.",
                "operationId": "2696cb94b790c1a4b3622e04592df080",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Halo HR, apakah ada kelanjutan wawancara?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pesan berhasil terkirim",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Pesan berhasil terkirim.",
                                    "data": {
                                        "id": 2,
                                        "application_id": 10,
                                        "sender_id": 1,
                                        "message": "Halo HR, apakah ada kelanjutan wawancara?",
                                        "created_at": "2026-08-21T22:10:00.000000Z"
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum Bearer Token to access protected API endpoints.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}